+++ e2e-tests/config/helpers/after.scenario.ts
@@ -7,7 +7,12 @@ interface World {
     'attach': ((arg1: string | Buffer, arg2: string) => void);
 }
 
-// Because @types/cucumber is not compatible with Cucumber 3 I've extended it for now
+/**
+ * In CucumberJS version 3 the `HookScenarioResult` has been brought down from an object with all the
+ * feature / scenario / step data in it to a bare minimum.
+ * Because `@types/cucumber` is not compatible with CucumberJS 3 I've extended the `HookScenarioResult`
+ * with the new interface
+ */
 interface TestCase extends HookScenarioResult {
     sourceLocation: ({
         uri: string;