+++ e2e-tests/config/protractor.shared.conf.js
@@ -1,102 +1,52 @@
 'use strict';
 const argv = require('yargs').argv;
-const fs = require('fs-extra');
 const path = require('path');
-const multiCucumberHTLMReporter = require('multiple-cucumber-html-reporter');
-let instanceData;
-const resultsFolder = '.tmp/results';
-const jsonOutputFolder = './.tmp/json-output-folder';
-const cucumberReportName = `results`;
 
 
-exports.config = sharedConfig();
+exports.config = {
+    /**
+     * Protractor specific
+     */
+    allScriptsTimeout: 11000,
+    disableChecks: true,
 
-function sharedConfig() {
-    const config = {
-        framework: 'custom',
-        frameworkPath: require.resolve('protractor-cucumber-framework'),
-        cucumberOpts: {
-            compiler: "ts:ts-node/register",
-            require: [
-                path.resolve(process.cwd(), './e2e-tests/**/after.scenario.ts'),
-                path.resolve(process.cwd(), './e2e-tests/**/cucumber.config.ts'),
-                path.resolve(process.cwd(), './e2e-tests/**/*.steps.ts')
-            ],
-            format: `json:${resultsFolder}/${cucumberReportName}.json`,
-            tags: ''
-        },
-        specs: getFeatureFiles(),
+    /**
+     * CucumberJS specific
+     */
+    framework: 'custom',
+    frameworkPath: require.resolve('protractor-cucumber-framework'),
+    cucumberOpts: {
+        compiler: "ts:ts-node/register",
+        require: [
+            path.resolve(process.cwd(), './e2e-tests/**/after.scenario.ts'),
+            path.resolve(process.cwd(), './e2e-tests/**/cucumber.config.ts'),
+            path.resolve(process.cwd(), './e2e-tests/**/*.steps.ts')
+        ],
+        format: 'json:.tmp/results.json',
+        tags: ''
+    },
+    specs: getFeatureFiles(),
 
-        beforeLaunch: () =>{
-            fs.ensureDirSync(resultsFolder);
-            fs.ensureDirSync(jsonOutputFolder);
-        },
+    /**
+     * From `protractor-cucumber-framework`, allows cucumber to handle the 199
+     * exception and record it appropriately
+     */
+    ignoreUncaughtExceptions: true,
 
-        onPrepare: function () {
-            // place something here
-            return browser.getCapabilities()
-                .then((capabilities) => {
-                    browser.browserName = capabilities.get('browserName').toLowerCase();
-
-                    instanceData = {
-                        pid: process.pid,
-                        metadata: {
-                            browser: {
-                                name: browser.browserName,
-                                version: capabilities.get('version') || capabilities.get('browserVersion')
-        },
-                            device: 'MacBook Pro 15',
-                            deviceType: 'laptop',
-                            environment: 'local',
-                            platform: {
-                                name: 'OSX',
-                                version: '10.12.6'
-                            }
-                        }
-                    };
-                });
-        },
-        onCleanUp: function () {
-            if (fs.existsSync(`./${resultsFolder}/results.${instanceData.pid}.json`)) {
-                const jsonReport = fs.readJsonSync(`./${resultsFolder}/${cucumberReportName}.${instanceData.pid}.json`);
-
-                jsonReport.map((singleReport) => {
-                    const featureName = singleReport.name.replace(/\s+/g, '_').replace(/\W/g, '').toLowerCase() || 'noName';
-                    const filePath = path.join(jsonOutputFolder, `${featureName}.${browser.browserName}_${Date.now()}.json`);
-
-                    // Remove a previous file if it exists to prevent double reports of 1 feature + browser execution
-                    fs.readdirSync(jsonOutputFolder)
-                        .filter((file) => file.match(new RegExp(`${featureName}.${browser.browserName}`, 'ig')))
-                        .forEach((file) => fs.removeSync(`${jsonOutputFolder}/${file}`));
-
-                    singleReport.metadata = instanceData.metadata;
-                    fs.writeJsonSync(filePath, JSON.parse(`[${JSON.stringify(singleReport)}]`), { spaces: 2 });
-                });
-            } else {
-                console.warn(`File: './${resultsFolder}/${cucumberReportName}.${instanceData.pid}.json' is not present.`);
-            }
-        },
-
-        afterLaunch: function(){
-            fs.removeSync(resultsFolder);
-
-            multiCucumberHTLMReporter.generate({
-                openReportInBrowser: true,
-                saveCollectedJSON: true,
-                jsonDir: '.tmp/json-output-folder',
-                reportPath: './.tmp/report/'
-            });
-        },
-
-        allScriptsTimeout: 11000,
-        disableChecks: true,
-
-        // From `protractor-cucumber-framework`, allows cucumber to handle the 199 exception and record it appropriately
-        ignoreUncaughtExceptions: true
-    };
-
-    return config;
-}
+    /**
+     * The new reporting plugin
+     */
+    plugins: [{
+        package: 'protractor-multiple-cucumber-html-reporter-plugin',
+        options: {
+            automaticallyGenerateReport: true,
+            metadataKey: 'deviceProperties',
+            removeExistingJsonReportFile: true,
+            removeOriginalJsonReportFile: true,
+            saveCollectedJSON: true
+        }
+    }]
+};
 
 /**
  * Get the featurefiles that need to be run based on an command line flag that is passed, if nothing is passed all the