+++ CHANGELOG.md
@@ -1,3 +1,52 @@
+<a name="3.0.0"></a>
+# 3.0.0
+
+## Changes to the boilerplate
+
+- Cleaned up the `protractor.shared.conf.js`:
+    - remove the `cucumberOpts.format: 'pretty'`, because it was removed by CucumberJS and added to `cucumberOpts.format: 'json:.tmp/results.json'` to nicely integrate with the new report plugin
+    - added `protractor-multiple-cucumber-html-reporter-plugin` to simplify generating and using CucumberJS reports
+- Remove the `reporter.ts`-hook because CucumberJS 3 removed the `registerListener` which was used to manipulate and save the JSON-report file
+- In CucumberJS version 3 the `HookScenarioResult` has been brought down from an object with all the feature / scenario / step data in it to a bare minimum. Because `@types/cucumber` is not compatible with CucumberJS 3 I've extended the `HookScenarioResult` in the `after.scenario.ts` with the new interface
+
+```js
+interface TestCase extends HookScenarioResult {
+    sourceLocation: ({
+        uri: string;
+        line: number
+    });
+    result: ({
+        duration: number;
+        status: string
+    });
+}
+```
+
+## Dependencies:
+
+### Added
+- `protractor-multiple-cucumber-html-reporter-plugin`
+- `protractor-flake`
+
+### Update
+- `protractor-cucumber-framework` to 4.0.7
+- `cucumber` to 3.0.1
+- `@types/chai` to 4.0.4
+- `@types/chai-as-promised` to 7.1.0
+- `@types/cucumber` to 2.0.3
+- `@types/fs-extra` to 4.0.2
+- `chai` to 4.1.2
+- `chai-as-promised` to 7.1.1
+- `fs-extra` to 4.0.1
+- `ts-node` to 3.3.0
+- `tslint` to 5.7.0
+- `typescript` to 2.5.2
+- `yargs` to 8.0.2
+
+### Remove
+- `jsonfile`
+- `multiple-cucumber-html-reporter`
+
 <a name="2.0.0"></a>
 # 2.0.0
 
@@ -39,7 +88,7 @@ this.After(function (scenarioResult:HookScenarioResult): Promise<void>{
 
 ### Removed `isFailed()`
 The `isFailed()` method has been removed from the `scenarioResult: HookScenarioResult` in comparison to `scenario: HookScenario`
- 
+
 ```javascript
 // 1.3.x
 this.After(function (scenario: HookScenario): void {
@@ -51,9 +100,9 @@ this.After(function (scenario: HookScenario): void {
 this.After(function (scenarioResult: HookScenarioResult): void {
     // logs scenario.status = '{string} status'
     console.log('scenarioResult.status = ', scenarioResult.status);
-    
+
     // Changed to
-    console.log('scenario failed = ', scenarioResult.status === 'failed');  
+    console.log('scenario failed = ', scenarioResult.status === 'failed');
 });
 ```